Add codes for 64bit operation

+++ src/strfunc.h
@@ -1,8 +1,9 @@
 #pragma once
 
-#include "osforteutype.h"
+#include "type.h"
+#include "math64.h"
 
 
 UINT32 strlen(UINT8* str);
-UINT8* itoa(UINT32 num);
+UINT8* itoa(UINT64 num);
 BOOL cmpstr(UINT8* str1, UINT8* str2);
